+++ ui/src/common/artist.js
