+++ ui/src/common/artist.test.js
